Fix compile

+++ zeus-java/zeus-batch/src/main/scala/io/github/zeus/batch/format/simple/FieldHelper.scala
@@ -13,8 +13,8 @@ object FieldHelper {
     def serialize(values: Iterator[Any], output: ColumnOutputStream): Int = {
       filedType match {
         case BOOL => BooleanColumnSerde.serialize(values.map(_.asInstanceOf[Boolean]), output)
-        case BYTE => ByteColumnSerde.serialize(values.map(_.asInstanceOf[Byte]), output)
-        case FLOAT => FloatColumnSerde.serialize(values.map(_.asInstanceOf[Float]), output)
+        case INT8=> ByteColumnSerde.serialize(values.map(_.asInstanceOf[Byte]), output)
+        case INT16=> FloatColumnSerde.serialize(values.map(_.asInstanceOf[Float]), output)
         case INT32 => IntColumnSerde.serialize(values.map(_.asInstanceOf[Int]), output)
         case INT64 => LongColumnSerde.serialize(values.map(_.asInstanceOf[Long]), output)
         case STRING => StringColumnSerde.serialize(values.map(_.asInstanceOf[String]), output)
